Extract gql logic from Questionnaire/Page
